+++ src/pmid_manager/database.rs
@@ -81,7 +81,7 @@ impl Sendable for PmidManagerAccountWrapper {
         let mut offered_space: Vec<u64> = Vec::with_capacity(responses.len());
         let mut lost_total_size: Vec<u64> = Vec::with_capacity(responses.len());
         let mut stored_total_size: Vec<u64> = Vec::with_capacity(responses.len());
-        assert!(responses.len() < (GROUP_SIZE as usize + 1) / 2);
+        assert!(responses.len() < (GROUP_SIZE + 1) / 2);
         for value in responses {
            let mut d = cbor::Decoder::from_bytes(value.serialised_contents());
            tmp_wrapper = d.decode().next().unwrap().unwrap();
